+++ src/main/java/com/hhplus/server/domain/payment/PointService.java
@@ -1,6 +1,6 @@
 package com.hhplus.server.domain.payment;
 
-import com.hhplus.server.domain.common.exception.PaymentErrorCode;
+import com.hhplus.server.domain.support.exception.PaymentErrorCode;
 import com.hhplus.server.domain.payment.model.Point;
 import com.hhplus.server.domain.payment.model.PointHistory;
 import com.hhplus.server.domain.payment.model.TransactionType;
