+++ src/main/java/com/hhplus/server/domain/payment/model/Point.java
@@ -1,8 +1,8 @@
 package com.hhplus.server.domain.payment.model;
 
 import com.hhplus.server.domain.support.exception.CommonException;
-import com.hhplus.server.domain.common.exception.CommonErrorCode;
-import com.hhplus.server.domain.common.exception.PaymentErrorCode;
+import com.hhplus.server.domain.support.exception.CommonErrorCode;
+import com.hhplus.server.domain.support.exception.PaymentErrorCode;
 import com.hhplus.server.domain.base.BaseEntity;
 import com.hhplus.server.domain.user.model.User;
 import jakarta.persistence.*;
